+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.1.11
+
+- Fixed an issue with `generatedNamespaces` during session proposal
+- Small enhancements in example wallet/dapp.
+- Minor changes and bug fixe
+
 ## 2.1.10
 
 - License change